pyproject.toml
[project]
name = "sql-guard"
version = "0.1.0"
description = "Trust and observability layer for text-to-SQL agents"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=0.10",
    "streamlit>=1.35",
    "typer>=0.12",
    "pydantic>=2.0",
    "ollama>=0.2",
    "pyaml>=23.0",
    "rich>=13.0",
    "fastapi>=0.111",
    "uvicorn>=0.30",
    "httpx>=0.27",
]

[project.scripts]
sql-guard = "sql_guard.cli:app"

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov", "ruff", "mypy"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
